move runtoolswithstreaming outside of mcp.ts
